Refactor: moved stuff from restaurant controller to cartcontroller
