+++ FoodApp/Views/Restaurants/Menu.cshtml
@@ -39,7 +39,7 @@
             @if (signInManager.IsSignedIn(User) && User.IsInRole("User"))
             {
                 <td>
-                    <a asp-action="AddToCart" asp-route-id="@item.Id">Add to cart</a>
+                    <a asp-controller="Cart" asp-action="AddToCart" asp-route-id="@item.Id">Add to cart</a>
 
                 </td>
             }
@@ -54,7 +54,7 @@
 <div>
     @if (signInManager.IsSignedIn(User) && User.IsInRole("User"))
             {
-    <a asp-action="Cart">Go to cart</a>
+    <a asp-controller="Cart" asp-action="Cart">Go to cart</a>
             }
         <a asp-action="Index">Back to List</a>
 </div>
