+++ FoodApp/Views/Restaurants/OrderDetails.cshtml
@@ -36,6 +36,7 @@
         }
     </tbody>
 </table>
+<h1>@Model.Client.Address</h1>
 
     <div class="row">
         <div class="col-md-4">
